+++ lib/Service/Gateway/SMS/Provider/ClickSendConfig.php
@@ -29,6 +29,10 @@ use OCA\TwoFactorGateway\Exception\ConfigurationException;
 use OCP\IConfig;
 
 class ClickSendConfig implements IProviderConfig {
+	private const expected = [
+		'clicksend_user',
+		'clicksend_apikey',
+	];
 
 	/** @var IConfig */
 	private $config;
@@ -38,7 +42,7 @@ class ClickSendConfig implements IProviderConfig {
 	}
 
 	private function getOrFail(string $key): string {
-		$val = $this->config->getAppValue(Application::APP_NAME, $key, null);
+		$val = $this->config->getAppValue(Application::APP_ID, $key, null);
 		if (is_null($val)) {
 			throw new ConfigurationException();
 		}
@@ -50,7 +54,7 @@ class ClickSendConfig implements IProviderConfig {
 	}
 
 	public function setUser(string $user) {
-		$this->config->setAppValue(Application::APP_NAME, 'clicksend_user', $user);
+		$this->config->setAppValue(Application::APP_ID, 'clicksend_user', $user);
 	}
 
 	public function getApiKey(): string {
@@ -58,15 +62,17 @@ class ClickSendConfig implements IProviderConfig {
 	}
 
 	public function setApiKey(string $password) {
-		$this->config->setAppValue(Application::APP_NAME, 'clicksend_apikey', $password);
+		$this->config->setAppValue(Application::APP_ID, 'clicksend_apikey', $password);
 	}
 
 	public function isComplete(): bool {
-		$set = $this->config->getAppKeys(Application::APP_NAME);
-		$expected = [
-			'clicksend_user',
-			'clicksend_apikey',
-		];
-		return count(array_intersect($set, $expected)) === count($expected);
+		$set = $this->config->getAppKeys(Application::APP_ID);
+		return count(array_intersect($set, self::expected)) === count(self::expected);
+	}
+
+	public function remove() {
+		foreach (self::expected as $key) {
+			$this->config->deleteAppValue(Application::APP_ID, $key);
+		}
 	}
 }