+++ lib/Service/Gateway/SMS/Provider/ClickatellCentralConfig.php
@@ -30,8 +30,7 @@ use OCA\TwoFactorGateway\Exception\ConfigurationException;
 use OCP\IConfig;
 
 class ClickatellCentralConfig implements IProviderConfig {
-
-	const expected = [
+	private const expected = [
 		'clickatell_central_api',
 		'clickatell_central_user',
 		'clickatell_central_password',
@@ -45,7 +44,7 @@ class ClickatellCentralConfig implements IProviderConfig {
 	}
 
 	private function getOrFail(string $key): string {
-		$val = $this->config->getAppValue(Application::APP_NAME, $key, null);
+		$val = $this->config->getAppValue(Application::APP_ID, $key, null);
 		if (is_null($val)) {
 			throw new ConfigurationException();
 		}
@@ -57,7 +56,7 @@ class ClickatellCentralConfig implements IProviderConfig {
 	}
 
 	public function setApi(string $api) {
-		$this->config->setAppValue(Application::APP_NAME, 'clickatell_central_api', $api);
+		$this->config->setAppValue(Application::APP_ID, 'clickatell_central_api', $api);
 	}
 
 	public function getUser(): string {
@@ -65,7 +64,7 @@ class ClickatellCentralConfig implements IProviderConfig {
 	}
 
 	public function setUser(string $user) {
-		$this->config->setAppValue(Application::APP_NAME, 'clickatell_central_user', $user);
+		$this->config->setAppValue(Application::APP_ID, 'clickatell_central_user', $user);
 	}
 
 	public function getPassword(): string {
@@ -73,17 +72,17 @@ class ClickatellCentralConfig implements IProviderConfig {
 	}
 
 	public function setPassword(string $password) {
-		$this->config->setAppValue(Application::APP_NAME, 'clickatell_central_password', $password);
+		$this->config->setAppValue(Application::APP_ID, 'clickatell_central_password', $password);
 	}
 
 	public function isComplete(): bool {
-		$set = $this->config->getAppKeys(Application::APP_NAME);
+		$set = $this->config->getAppKeys(Application::APP_ID);
 		return count(array_intersect($set, self::expected)) === count(self::expected);
 	}
 
 	public function remove() {
-		foreach(self::expected as $key) {
-			$this->config->deleteAppValue(Application::APP_NAME, $key);
+		foreach (self::expected as $key) {
+			$this->config->deleteAppValue(Application::APP_ID, $key);
 		}
 	}
 }