+++ lib/Service/Gateway/SMS/Provider/ProviderFactory.php
@@ -37,10 +37,14 @@ class ProviderFactory {
 
 	public function getProvider(string $id): IProvider {
 		switch ($id) {
+			case SipGate::PROVIDER_ID:
+				return $this->container->query(SipGate::class);
 			case PuzzelSMS::PROVIDER_ID:
 				return $this->container->query(PuzzelSMS::class);
 			case PlaySMS::PROVIDER_ID:
 				return $this->container->query(PlaySMS::class);
+			case SMSGlobal::PROVIDER_ID:
+				return $this->container->query(SMSGlobal::class);
 			case WebSms::PROVIDER_ID:
 				return $this->container->query(WebSms::class);
 			case ClockworkSMS::PROVIDER_ID:
@@ -49,6 +53,8 @@ class ProviderFactory {
 				return $this->container->query(EcallSMS::class);
 			case VoipMs::PROVIDER_ID:
 				return $this->container->query(VoipMs::class);
+			case Voipbuster::PROVIDER_ID:
+				return $this->container->query(Voipbuster::class);
 			case HuaweiE3531::PROVIDER_ID:
 				return $this->container->query(HuaweiE3531::class);
 			case Sms77Io::PROVIDER_ID:
@@ -63,6 +69,8 @@ class ProviderFactory {
 				return $this->container->query(ClickatellPortal::class);
 			case ClickSend::PROVIDER_ID:
 				return $this->container->query(ClickSend::class);
+			case SerwerSMS::PROVIDER_ID:
+				return $this->container->query(SerwerSMS::class);
 			default:
 				throw new InvalidSmsProviderException("Provider <$id> does not exist");
 		}