+++ lib/Service/Gateway/SMS/Provider/WebSmsConfig.php
@@ -29,8 +29,7 @@ use OCA\TwoFactorGateway\Exception\ConfigurationException;
 use OCP\IConfig;
 
 class WebSmsConfig implements IProviderConfig {
-
-	const expected = [
+	private const expected = [
 		'websms_de_user',
 		'websms_de_password',
 	];
@@ -43,7 +42,7 @@ class WebSmsConfig implements IProviderConfig {
 	}
 
 	private function getOrFail(string $key): string {
-		$val = $this->config->getAppValue(Application::APP_NAME, $key, null);
+		$val = $this->config->getAppValue(Application::APP_ID, $key, null);
 		if (is_null($val)) {
 			throw new ConfigurationException();
 		}
@@ -55,7 +54,7 @@ class WebSmsConfig implements IProviderConfig {
 	}
 
 	public function setUser(string $user) {
-		$this->config->setAppValue(Application::APP_NAME, 'websms_de_user', $user);
+		$this->config->setAppValue(Application::APP_ID, 'websms_de_user', $user);
 	}
 
 	public function getPassword(): string {
@@ -63,17 +62,17 @@ class WebSmsConfig implements IProviderConfig {
 	}
 
 	public function setPassword(string $password) {
-		$this->config->setAppValue(Application::APP_NAME, 'websms_de_password', $password);
+		$this->config->setAppValue(Application::APP_ID, 'websms_de_password', $password);
 	}
 
 	public function isComplete(): bool {
-		$set = $this->config->getAppKeys(Application::APP_NAME);
+		$set = $this->config->getAppKeys(Application::APP_ID);
 		return count(array_intersect($set, self::expected)) === count(self::expected);
 	}
 
 	public function remove() {
-		foreach(self::expected as $key) {
-			$this->config->deleteAppValue(Application::APP_NAME, $key);
+		foreach (self::expected as $key) {
+			$this->config->deleteAppValue(Application::APP_ID, $key);
 		}
 	}
 }